+++ src/app/app.scss
@@ -1,2 +1,2 @@
 @import '~react-components/styles/index';
-@import 'component/Labels/LabelsColor';
+/*@import 'components/Labels/LabelsColor.scss';*/
