+++ src/app/components/Labels/LabelColors.scss
