+++ src/app/components/Labels/LabelSortableItem.js
